+++ Frontend-v1-Original/components/ssBribeCreate/queries.ts
@@ -1,34 +1,30 @@
-import { useQuery } from "@tanstack/react-query";
-import { useAccount } from "wagmi";
+import { useMemo } from "react";
 
 import {
-  getBaseAssetsWithInfo,
-  getPairs,
-  usePairsData,
+  useBaseAssetWithInfo,
+  usePairsWithGauges,
 } from "../../lib/global/queries";
-import { NATIVE_TOKEN, QUERY_KEYS } from "../../stores/constants/constants";
+import { NATIVE_TOKEN } from "../../stores/constants/constants";
 import { hasGauge } from "../../stores/types/types";
 
 export const useGauges = () => {
-  const { data: pairsData } = usePairsData();
-  return useQuery({
-    queryKey: [QUERY_KEYS.GAUGES, pairsData],
-    queryFn: () => getPairs(pairsData),
-    enabled: !!pairsData,
-    select: (pairs) =>
-      pairs.filter(hasGauge).filter((gauge) => gauge.isAliveGauge),
-  });
+  const query = usePairsWithGauges();
+  return {
+    ...query,
+    data: useMemo(() => {
+      return query.data?.filter(hasGauge).filter((gauge) => gauge.isAliveGauge);
+    }, [query.data]),
+  };
 };
 
 export const useBaseAssetWithInfoNoNative = () => {
-  const { address } = useAccount();
-  return useQuery({
-    queryKey: [QUERY_KEYS.BASE_ASSET_INFO_NO_NATIVE, address],
-    queryFn: () => getBaseAssetsWithInfo(address),
-    select: (baseAssets) => {
-      return baseAssets.filter((option) => {
+  const query = useBaseAssetWithInfo();
+  return {
+    ...query,
+    data: useMemo(() => {
+      return query.data?.filter((option) => {
         return option.address !== NATIVE_TOKEN.address;
       });
-    },
-  });
+    }, [query.data]),
+  };
 };
