+++ Frontend-v1-Original/components/ssLiquidityManage/queries.ts
@@ -1,31 +1,19 @@
-import { useQuery } from "@tanstack/react-query";
-import { useAccount } from "wagmi";
+import { useMemo } from "react";
 import BigNumber from "bignumber.js";
 
-import {
-  getPairsWithGauges,
-  usePairsWithoutGauges,
-} from "../../lib/global/queries";
-import { QUERY_KEYS } from "../../stores/constants/constants";
+import { usePairsWithGauges } from "../../lib/global/queries";
 
 export const usePairsWithGaugesOnlyWithBalance = () => {
-  const { address } = useAccount();
-  const { data: pairsWithoutGauges } = usePairsWithoutGauges();
-  return useQuery({
-    queryKey: [
-      QUERY_KEYS.PAIRS_WITH_GAUGES_AND_BALANCES,
-      address,
-      pairsWithoutGauges,
-    ],
-    queryFn: () => getPairsWithGauges(address!, pairsWithoutGauges!),
-    enabled: !!address && !!pairsWithoutGauges,
-    select: (pairs) => {
-      return pairs.filter((ppp) => {
+  const query = usePairsWithGauges();
+  return {
+    ...query,
+    data: useMemo(() => {
+      return query.data?.filter((ppp) => {
         return (
           (ppp.balance && BigNumber(ppp.balance).gt(0)) ||
           (ppp.gauge?.balance && BigNumber(ppp.gauge.balance).gt(0))
         );
       });
-    },
-  });
+    }, [query.data]),
+  };
 };
