+++ Frontend-v1-Original/stores/constants/actions.ts
@@ -116,8 +116,8 @@ export const VOTE_RETURNED = "VOTE_RETURNED";
 export const GET_VEST_VOTES = "GET_VEST_VOTES";
 export const VEST_VOTES_RETURNED = "VEST_VOTES_RETURNED";
 
-export const CREATE_BRIBE = "CREATE_BRIBE";
-export const BRIBE_CREATED = "BRIBE_CREATED";
+// export const CREATE_BRIBE = "CREATE_BRIBE";
+// export const BRIBE_CREATED = "BRIBE_CREATED";
 
 // export const GET_VEST_BALANCES = "GET_VEST_BALANCES";
 // export const VEST_BALANCES_RETURNED = "VEST_BALANCES_RETURNED";
@@ -143,10 +143,10 @@ export const CLAIM_VE_DIST = "CLAIM_VE_DIST";
 export const CLAIM_VE_DIST_RETURNED = "CLAIM_VE_DIST_RETURNED";
 
 //WHITELIST
-export const SEARCH_WHITELIST = "SEARCH_WHITELIST";
-export const SEARCH_WHITELIST_RETURNED = "SEARCH_WHITELIST_RETURNED";
-export const WHITELIST_TOKEN = "WHITELIST_TOKEN";
-export const WHITELIST_TOKEN_RETURNED = "WHITELIST_TOKEN_RETURNED";
+// export const SEARCH_WHITELIST = "SEARCH_WHITELIST";
+// export const SEARCH_WHITELIST_RETURNED = "SEARCH_WHITELIST_RETURNED";
+// export const WHITELIST_TOKEN = "WHITELIST_TOKEN";
+// export const WHITELIST_TOKEN_RETURNED = "WHITELIST_TOKEN_RETURNED";
 
 //AUTO BRIBE
 // export const BRIBE_AUTO_BRIBE = "BRIBE_AUTO_BRIBE";
